# docs/swagger.yaml
basePath: /
definitions:
  github.com_miprokop_fication_internal_http-server_handlers.errorResponse:
    properties:
      message:
        type: string
    type: object
...
  models.Event:
    properties:
      created_by:
        $ref: '#/definitions/models.Staff'
      created_by_id:
        type: string
      creation_date:
        type: string
      description:
        type: string
      end_date:
        type: string
      event_status:
        type: string
      event_type:
        type: string
      id:
        type: string
      image_path:
        type: string
      name:
        type: string
      organization_id:
        type: string
      staff:
        items:
          $ref: '#/definitions/models.StaffEvents'
        type: array
      steps:
        items:
          $ref: '#/definitions/models.Step'
        type: array
    type: object
  models.Organization:
    properties:
      id:
        type: string
      image:
        type: string
      name:
        type: string
      positions:
        items:
          $ref: '#/definitions/models.Position'
        type: array
      teams:
        items:
          $ref: '#/definitions/models.Team'
        type: array
      types:
        items:
          $ref: '#/definitions/models.OrganizationType'
        type: array
      website_url:
        type: string
    type: object
  models.OrganizationType:
    properties:
      id:
        type: string
      name:
        type: string
      organization:
        items:
          $ref: '#/definitions/models.Organization'
        type: array
    type: object
  models.Permission:
    properties:
      granted_by:
        type: string
      permission:
        type: string
      position:
        $ref: '#/definitions/models.Position'
      position_id:
        type: string
    type: object
  models.Position:
    properties:
      company_id:
        type: string
      id:
        type: string
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/models.Permission'
        type: array
    type: object
  models.Prize:
    properties:
      count:
        type: integer
      created_by:
        type: string
      creation_date:
        type: string
      current_count:
        type: integer
      data:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      prizes:
        items:
          $ref: '#/definitions/models.StaffPrize'
        type: array
      staff:
        $ref: '#/definitions/models.Staff'
      status:
        type: string
      step:
        $ref: '#/definitions/models.Step'
      step_id:
        type: string
      type:
        type: string
    type: object
  models.Staff:
    properties:
      additional_info:
        type: string
      background_color:
        type: string
      current_image:
        type: string
      email:
        type: string
      events:
        items:
          $ref: '#/definitions/models.StaffEvents'
        type: array
      first_name:
        type: string
      id:
        type: string
      images:
        items:
          $ref: '#/definitions/models.StaffImage'
        type: array
      last_name:
        type: string
      organization_id:
        type: string
      position:
        $ref: '#/definitions/models.Position'
      position_id:
        type: string
      prizes:
        items:
          $ref: '#/definitions/models.StaffPrize'
        type: array
      sex:
        type: string
      team:
        $ref: '#/definitions/models.Team'
      team_id:
        type: string
      text_color:
        type: string
    type: object
  models.StaffEvents:
    properties:
      event:
        $ref: '#/definitions/models.Event'
      event_id:
        type: string
      id:
        type: string
      staff:
        $ref: '#/definitions/models.Staff'
      staff_id:
        type: string
      staff_role:
        type: string
      status:
        type: string
    type: object
...
  models.StaffImage:
    properties:
      id:
        type: string
      image_path:
        type: string
    type: object
...
  models.StaffLogin:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.StaffPrize:
    properties:
      id:
        type: string
      prize:
        $ref: '#/definitions/models.Prize'
      prize_id:
        type: string
      staff:
        $ref: '#/definitions/models.Staff'
      staff_id:
        type: string
    type: object
  models.Step:
    properties:
      active_staff:
        items:
          $ref: '#/definitions/models.Staff'
        type: array
      creation_date:
        type: string
      description:
        type: string
      end_date:
        type: string
      event_id:
        type: string
      id:
        type: string
      images:
        items:
          $ref: '#/definitions/models.StepImage'
        type: array
      level:
        type: integer
      max_score:
        type: integer
      name:
        type: string
      prizes:
        items:
          $ref: '#/definitions/models.Prize'
        type: array
      status:
        type: string
      task:
        type: string
    type: object
  models.StepImage:
    properties:
      id:
        type: string
      image_url:
        type: string
      step:
        $ref: '#/definitions/models.Step'
      step_id:
        type: string
    type: object
  models.Team:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      organization:
        $ref: '#/definitions/models.Organization'
      organization_id:
        type: string
      staff:
        items:
          $ref: '#/definitions/models.Staff'
        type: array
    type: object
host: localhost:8082
info:
  contact: {}
  description: API Server for challenge service
  title: Acheer API
  version: "1.0"
paths:
...
  /sign-in:
    post:
      consumes:
      - application/json
      description: |-
        sign in staff to get token
        token is used in authorization
      operationId: sign-in-staff
      parameters:
      - description: staff account log in info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.StaffLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
      summary: signIn
      tags:
      - auth
  /sign-up:
    post:
      consumes:
      - application/json
      description: |-
        create models.Staff
        if no organization_id in input
        push staff do default org
        do the same stuff with position_id and team_id if organization_id is empty
        also in registering you can pass image
      operationId: create-staff
      parameters:
      - description: staff account info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Staff'
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
      summary: SignUp
      tags:
      - auth
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"